use new test fixture
