+++ src/main/scala/ai/privado/languageEngine/csharp/processor/CSharpProcessor.scala
@@ -117,7 +117,7 @@ class CSharpProcessor(
       .withInputPath(sourceRepoLocation)
       .withOutputPath(cpgOutputPath)
       .withIgnoredFilesRegex(excludeFileRegex)
-      .withDownloadDependencies(!config.skipDownloadDependencies)
+      .withDownloadDependencies(Try(!config.skipDownloadDependencies).getOrElse(false))
 
     val xtocpg = new CSharpSrc2Cpg().createCpg(cpgconfig).map { cpg =>
       println(
